Edit structure
