+++ src/internal/type-extensions.ts
@@ -1,16 +1,27 @@
 import type {hethers} from "@hashgraph/hethers";
 import {HardhatConfig} from "hardhat/types/config";
 import {HardhatRuntimeEnvironment, Network} from "hardhat/types/runtime";
-import { SignerWithAddress } from "../signers";
-import { FactoryOptions } from "../types";
-import { Artifact } from "hardhat/types";
+import {SignerWithAddress} from "./signers";
+import {Artifact} from "hardhat/types";
 
-// import type {
-//   FactoryOptions as FactoryOptionsT,
-//   getContractFactory as getContractFactoryT,
-//   HardhatEthersHelpers,
-//   Libraries as LibrariesT,
-// } from "../types";
+export interface Libraries {
+    [libraryName: string]: string;
+}
+
+export interface FactoryOptions {
+    signer?: hethers.Signer;
+    libraries?: Libraries;
+}
+
+export declare function getContractFactory(
+    name: string,
+    signerOrOptions?: hethers.Signer | FactoryOptions
+): Promise<hethers.ContractFactory>;
+export declare function getContractFactory(
+    abi: any[],
+    bytecode: hethers.utils.BytesLike,
+    signer?: hethers.Signer
+): Promise<hethers.ContractFactory>;
 
 export interface HederaAccount {
     account?: string;
@@ -51,38 +62,41 @@ interface HederaNetworkInterface extends Network {
 }
 
 type HethersT = typeof hethers;
+
 interface HethersTExtended extends HethersT {
     provider: any,
-    getSigners( hre: HederaHardhatRuntimeEnvironment ): Promise<SignerWithAddress[]>;
+
+    getSigners(hre: HederaHardhatRuntimeEnvironment): Promise<SignerWithAddress[]>;
+
     getSigner(hre: HederaHardhatRuntimeEnvironment, identifier: any): Promise<SignerWithAddress>;
 
     getContractFactory(
-      hre: HederaHardhatRuntimeEnvironment,
-      nameOrAbi: string | any[],
-      bytecodeOrFactoryOptions?:
-        | (hethers.Signer | FactoryOptions)
-        | hethers.utils.BytesLike,
-      signer?: hethers.Signer
+        hre: HederaHardhatRuntimeEnvironment,
+        nameOrAbi: string | any[],
+        bytecodeOrFactoryOptions?:
+            | (hethers.Signer | FactoryOptions)
+            | hethers.utils.BytesLike,
+        signer?: hethers.Signer
     ): Promise<hethers.ContractFactory>;
 
     getContractFactoryFromArtifact(
-      hre: HardhatRuntimeEnvironment,
-      artifact: Artifact,
-      signerOrOptions?: hethers.Signer | FactoryOptions
+        hre: HardhatRuntimeEnvironment,
+        artifact: Artifact,
+        signerOrOptions?: hethers.Signer | FactoryOptions
     ): Promise<hethers.ContractFactory>;
 
     getContractAt(
-      hre: HederaHardhatRuntimeEnvironment,
-      nameOrAbi: string | any[],
-      address: string,
-      signer?: hethers.Signer
+        hre: HederaHardhatRuntimeEnvironment,
+        nameOrAbi: string | any[],
+        address: string,
+        signer?: hethers.Signer
     ): Promise<hethers.ContractFactory>;
 
     getContractAtFromArtifact(
-      hre: HederaHardhatRuntimeEnvironment,
-      artifact: Artifact,
-      address: string,
-      signer?: hethers.Signer
+        hre: HederaHardhatRuntimeEnvironment,
+        artifact: Artifact,
+        address: string,
+        signer?: hethers.Signer
     ): Promise<hethers.ContractFactory>;
 }
